grouping: reworking for clarity and comprehensivness
Mapping internally categories and key values of netbox fields makes the
configuration less intiuitive and fragile.

This rework gives full control of how - from where exactly and in what
semantical constellation - to to consume the netbox data to the user of
the dynamic inventory. It's more expressive, but this comes bolsters
also expedited clarity.